requirements.txt
annotated-types==0.6.0
APScheduler==3.6.3
cachetools==4.2.2
Cerberus==1.3.5
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
platformdirs==4.1.0
plette==0.4.4
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tomlkit==0.12.3
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
yarg==0.1.9
